Fix warnings, create header files to relax warning settings for use with external includes

+++ source/gloperate-assimp/source/AssimpSceneLoader.cpp
@@ -5,12 +5,16 @@
 #include <iostream>
 #include <string>
 
+#include <gloperate/ext-includes-begin.h>
+
 #include <glm/glm.hpp>
 
 #include <assimp/scene.h>
 #include <assimp/cimport.h>
 #include <assimp/types.h>
 #include <assimp/postprocess.h>
+
+#include <gloperate/ext-includes-end.h>
 
 #include <gloperate/primitives/PolygonalGeometry.h>
 #include <gloperate/primitives/Scene.h>
@@ -145,7 +149,7 @@ Scene * AssimpSceneLoader::convertScene(const aiScene * scene) const
         sceneOut->meshes().push_back(convertGeometry(scene->mMeshes[i]));
     }
 
-    for (size_t i = 0; i < scene->mNumMaterials; ++i)
+    for (unsigned int i = 0; i < scene->mNumMaterials; ++i)
     {
         aiString filename;
         // only fetch texture with index 0
